moved scripts to new spynnaker_integration_tests
