+++ p8_integration_tests/quick_test/test_auto_pause_and_resume_tests/test_shorter_run/test_synfire_low_sdram_low_run.py
@@ -20,7 +20,7 @@ from testfixtures import LogCapture
 import spynnaker.plot_utils as plot_utils
 import spynnaker.spike_checker as spike_checker
 from spinnaker_testbase import BaseTestCase
-from p8_integration_tests.scripts.synfire_run import SynfireRunner
+from spynnaker_integration_tests.scripts.synfire_run import SynfireRunner
 
 n_neurons = 200  # number of neurons in each population
 runtime = 3000
