+++ p8_integration_tests/quick_test/test_change_neuron_parameters_between_runs/test_synfire_poisson_if_curr_exp_parameter.py
@@ -16,7 +16,7 @@ import numpy
 import spynnaker.plot_utils as plot_utils
 import spynnaker8 as p
 from spinnaker_testbase import BaseTestCase
-from p8_integration_tests.scripts.synfire_run import SynfireRunner
+from spynnaker_integration_tests.scripts.synfire_run import SynfireRunner
 
 n_neurons = 200  # number of neurons in each population
 neurons_per_core = n_neurons / 2
