+++ p8_integration_tests/quick_test/test_multi_call_examples/test_multi_extract.py
@@ -16,7 +16,7 @@
 import spynnaker8 as sim
 from spynnaker.pyNN.utilities import neo_compare
 from spinnaker_testbase import BaseTestCase
-from p8_integration_tests.scripts.checker import check_neuron_data
+from spynnaker_integration_tests.scripts import check_neuron_data
 
 n_neurons = 20  # number of neurons in each population
 neurons_per_core = n_neurons / 2
