+++ p8_integration_tests/quick_test/test_param_vertex_retrieval_from_board_after_1_run/test_synfire_200n_100pc_with_delays_all_recording.py
@@ -18,14 +18,13 @@
 """
 Synfirechain-like example
 """
-import os.path
 import matplotlib.pyplot as plt
 import numpy
 from pyNN.utility.plotting import Figure
 import spynnaker.spike_checker as spike_checker
 from spynnaker.spynnaker_plotting import SpynnakerPanel
 from spinnaker_testbase import BaseTestCase
-from p8_integration_tests.scripts.synfire_run import SynfireRunner
+from spynnaker_integration_tests.scripts.synfire_run import SynfireRunner
 
 n_neurons = 200  # number of neurons in each population
 delay = 17
