+++ p8_integration_tests/quick_test/test_param_vertex_retrieval_from_board_after_1_run/test_synfire_20n_20pc_no_delays_all_recording.py
@@ -14,14 +14,13 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-import os.path
 import matplotlib.pyplot as plt
 import numpy
 from pyNN.utility.plotting import Figure
 import spynnaker.spike_checker as spike_checker
 from spynnaker.spynnaker_plotting import SpynnakerPanel
 from spinnaker_testbase import BaseTestCase
-from p8_integration_tests.scripts.synfire_run import SynfireRunner
+from spynnaker_integration_tests.scripts.synfire_run import SynfireRunner
 """
 Synfirechain-like example
 """
