+++ spynnaker_integration_tests/scripts/checker.py
