+++ spynnaker_integration_tests/scripts/patternSpiker.py
