+++ spynnaker_integration_tests/scripts/synfire_npop_run.py
