+++ spynnaker_integration_tests/scripts/synfire_run.py
